requirements.txt
# Database Connectivity
psycopg2-binary>=2.9.0

# Database ORM
sqlalchemy>=1.4.0

# Data Processing
pandas>=1.5.0
pyarrow>=14.0.0

# HTTP Requests
requests>=2.28.0

# Configuration
pyyaml>=6.0

# Environment Variables
python-dotenv>=0.19.0

# Workflow Orchestration
prefect>=3.4.7

# Standard Library Modules (built-in, no installation required)
# logging - Used throughout for logging operations
# os - Environment and path operations
# sys - System path management
# string.Template - Template string substitution
# datetime - Date and time operations
# re - Regular expressions
# io - StringIO for data processing
# json - JSON data processing
# csv - CSV file operations
# collections.OrderedDict - Ordered dictionary operations
# random - Random data generation
# pathlib.Path - Path operations
